Correct casts and comments

+++ ubxgps.h
@@ -9,9 +9,9 @@
 #define UBX_VALUE_S8(x) UBX_VALUE_U8(x)
 
 // 16-bit values
-#define UBX_VALUE_U16(x) (uint8_t)(x & 0xFF), (uint8_t)((x >> 8) & 0xFF)
+#define UBX_VALUE_U16(x) (uint8_t)(((uint16_t)x) & 0xFF), (uint8_t)((((uint16_t)x) >> 8) & 0xFF)
 #define UBX_VALUE_S16(x) UBX_VALUE_U16(x)
 
 // 32-bit values
-#define UBX_VALUE_U32(x) (uint8_t)(x & 0xFF), (uint8_t)((x >> 8) & 0xFF), (uint8_t)((x >> 16) & 0xFF), (uint8_t)((x >> 24) & 0xFF)
+#define UBX_VALUE_U32(x) (uint8_t)(((uint32_t)x) & 0xFF), (uint8_t)((((uint32_t)x) >> 8) & 0xFF), (uint8_t)((((uint32_t)x) >> 16) & 0xFF), (uint8_t)((((uint32_t)x) >> 24) & 0xFF)
 #define UBX_VALUE_S32(x) UBX_VALUE_U32(x)
